+++ src/RcppExports.cpp
@@ -12,15 +12,13 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // do_boosting
-Rcpp::List do_boosting(arma::mat X, double precision, double alpha, double beta, double gamma, int max_resol, int num_each_dim, int num_second, double learn_rate, int min_obs, int nbins, double eta_subsample, double thresh_stop, int ntrees_wait, bool show_progress);
-RcppExport SEXP _boostPM_do_boosting(SEXP XSEXP, SEXP precisionSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP gammaSEXP, SEXP max_resolSEXP, SEXP num_each_dimSEXP, SEXP num_secondSEXP, SEXP learn_rateSEXP, SEXP min_obsSEXP, SEXP nbinsSEXP, SEXP eta_subsampleSEXP, SEXP thresh_stopSEXP, SEXP ntrees_waitSEXP, SEXP show_progressSEXP) {
+Rcpp::List do_boosting(arma::mat X, double prior_split_prob, double gamma, int max_resol, int num_each_dim, int num_second, double learn_rate, int min_obs, int nbins, double eta_subsample, double thresh_stop, int ntrees_wait, bool show_progress);
+RcppExport SEXP _boostPM_do_boosting(SEXP XSEXP, SEXP prior_split_probSEXP, SEXP gammaSEXP, SEXP max_resolSEXP, SEXP num_each_dimSEXP, SEXP num_secondSEXP, SEXP learn_rateSEXP, SEXP min_obsSEXP, SEXP nbinsSEXP, SEXP eta_subsampleSEXP, SEXP thresh_stopSEXP, SEXP ntrees_waitSEXP, SEXP show_progressSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< arma::mat >::type X(XSEXP);
-    Rcpp::traits::input_parameter< double >::type precision(precisionSEXP);
-    Rcpp::traits::input_parameter< double >::type alpha(alphaSEXP);
-    Rcpp::traits::input_parameter< double >::type beta(betaSEXP);
+    Rcpp::traits::input_parameter< double >::type prior_split_prob(prior_split_probSEXP);
     Rcpp::traits::input_parameter< double >::type gamma(gammaSEXP);
     Rcpp::traits::input_parameter< int >::type max_resol(max_resolSEXP);
     Rcpp::traits::input_parameter< int >::type num_each_dim(num_each_dimSEXP);
@@ -32,7 +30,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< double >::type thresh_stop(thresh_stopSEXP);
     Rcpp::traits::input_parameter< int >::type ntrees_wait(ntrees_waitSEXP);
     Rcpp::traits::input_parameter< bool >::type show_progress(show_progressSEXP);
-    rcpp_result_gen = Rcpp::wrap(do_boosting(X, precision, alpha, beta, gamma, max_resol, num_each_dim, num_second, learn_rate, min_obs, nbins, eta_subsample, thresh_stop, ntrees_wait, show_progress));
+    rcpp_result_gen = Rcpp::wrap(do_boosting(X, prior_split_prob, gamma, max_resol, num_each_dim, num_second, learn_rate, min_obs, nbins, eta_subsample, thresh_stop, ntrees_wait, show_progress));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -64,7 +62,7 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_boostPM_do_boosting", (DL_FUNC) &_boostPM_do_boosting, 15},
+    {"_boostPM_do_boosting", (DL_FUNC) &_boostPM_do_boosting, 13},
     {"_boostPM_simulation", (DL_FUNC) &_boostPM_simulation, 3},
     {"_boostPM_evaluate_log_density", (DL_FUNC) &_boostPM_evaluate_log_density, 3},
     {NULL, NULL, 0}